change __test__ folder structure for build
