+++ src/__tests__/__componentTests__/Account/AccountOrders.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import { testOrdersList } from '../../../../static/testOrdersList';
-import AccountOrders from '../../Account/AccountOrders';
+import AccountOrders from '../../../components/Account/AccountOrders';
 
 test('AccountOrders', () => {
   const component = renderer.create(<AccountOrders orders={testOrdersList} />);
