+++ src/__tests__/__componentTests__/Product/ProductAttributes.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { testUser } from '../../../../static/testData';
 import { testProductList } from '../../../../static/testProductList';
-import ProductAttributes from '../../Product/ProductAttributes';
+import ProductAttributes from '../../../components/Product/ProductAttributes';
 
 test('ProductAttributes', () => {
   const component = renderer.create(
