+++ src/__tests__/__componentTests__/_App/Header.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Layout from '../../_App/Layout';
 import { testUser } from '../../../../static/testData';
+import Header from '../../../components/_App/Header';
 
 jest.mock('next/router', () => ({
   events: {
@@ -23,8 +23,9 @@ jest.mock('next/router', () => ({
     };
   },
 }));
-test('Layout', () => {
-  const component = renderer.create(<Layout user={testUser} />);
+
+test('Header', () => {
+  const component = renderer.create(<Header user={testUser} />);
   const tree = component.toJSON();
 
   expect(tree).toMatchSnapshot();
