+++ src/__tests__/__pagesTests__/Create.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import CreateProduct from '../create';
+import CreateProduct from '../../pages/create';
 
 test('CreateProduct', () => {
   const component = renderer.create(<CreateProduct />);
